Profile Image updates for media queries
Suggested changes from Robertpm on Treehouse via Slack:

Okay, couple suggestions, Ryan.  First, target the containers of the
portfolio items, the class `.portfolio-piece`, and set the
`vertical-align` property to `top`.  This should fix the alignment
issue you talked about.  Then set a max-width property on the container
of all the portfolio items, the class `.portoflio`, and set that
max-width property to something like 950px.  That will keep those items
in a three column layout even on larger screens, as long as you add a
max-width property to the profile pic.  At present, that profile pic is
set with a percentage so it just keeps growing, and on larger screens,
it interferes with the portfolio items.  So if you set that profile
pic's max-width property to something like 400px or so, it should stop
doing that.

Hope that helps, bud!  Keep up the great work!